config: factor out common subscription test cases.
This patch takes the common Subscription test cases for gRPC/REST/inotify implementations and
moves them to a parameterized test.
