+++ test/common/config/BUILD
@@ -3,6 +3,7 @@ licenses(["notice"])  # Apache 2
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_test",
+    "envoy_cc_test_library",
     "envoy_package",
 )
 
@@ -11,8 +12,17 @@ envoy_package()
 envoy_cc_test(
     name = "filesystem_subscription_impl_test",
     srcs = ["filesystem_subscription_impl_test.cc"],
+    deps = [
+        ":filesystem_subscription_test_harness",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "filesystem_subscription_test_harness",
+    srcs = ["filesystem_subscription_test_harness.h"],
     external_deps = ["envoy_eds"],
     deps = [
+        ":subscription_test_harness",
         "//source/common/config:filesystem_subscription_lib",
         "//source/common/event:dispatcher_lib",
         "//test/mocks/config:config_mocks",
@@ -24,8 +34,17 @@ envoy_cc_test(
 envoy_cc_test(
     name = "grpc_subscription_impl_test",
     srcs = ["grpc_subscription_impl_test.cc"],
+    deps = [
+        ":grpc_subscription_test_harness",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "grpc_subscription_test_harness",
+    hdrs = ["grpc_subscription_test_harness.h"],
     external_deps = ["envoy_eds"],
     deps = [
+        ":subscription_test_harness",
         "//source/common/config:grpc_subscription_lib",
         "//test/mocks/config:config_mocks",
         "//test/mocks/event:event_mocks",
@@ -38,11 +57,20 @@ envoy_cc_test(
 envoy_cc_test(
     name = "http_subscription_impl_test",
     srcs = ["http_subscription_impl_test.cc"],
+    deps = [
+        ":http_subscription_test_harness",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "http_subscription_test_harness",
+    srcs = ["http_subscription_test_harness.h"],
     external_deps = [
         "envoy_eds",
         "protobuf",
     ],
     deps = [
+        ":subscription_test_harness",
         "//include/envoy/http:async_client_interface",
         "//source/common/common:utility_lib",
         "//source/common/config:http_subscription_lib",
@@ -54,4 +82,20 @@ envoy_cc_test(
         "//test/mocks/upstream:upstream_mocks",
         "//test/test_common:utility_lib",
     ],
+)
+
+envoy_cc_test(
+    name = "subscription_impl_test",
+    srcs = ["subscription_impl_test.cc"],
+    deps = [
+        ":filesystem_subscription_test_harness",
+        ":grpc_subscription_test_harness",
+        ":http_subscription_test_harness",
+        ":subscription_test_harness",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "subscription_test_harness",
+    srcs = ["subscription_test_harness.h"],
 )
